fix sender id collision between local and remote addresses
Switch the actor id type from usize to u64 throughout the workspace, then
reserve the high bit of the id space to disjoin local Address ids from
RemoteAddress ids so they can safely share a Recipient<M> hash space (e.g.
in ObserverSet). RemoteAddress::index now reverses session.index() into
the high bits and XORs with remote_actor_id in the low bits, which is
collision-free as long as log2(session) + log2(actor) < 63 — strictly
better than the previous (session << 32) + actor packing that broke for
session ids >= 2^31. The cached value is computed once at construction.

SenderIndex grows an is_remote() default that derives the answer from
index() >> 63, with a u64 base impl so wire-decode sites can validate
raw actor ids without knowing about REMOTE_FLAG. Encoders refuse to
serialize a Recipient that wraps a RemoteAddress, and decoders refuse
to construct one from a tagged wire id; both paths use new typed
EncodeRemoteAddress / DecodeRemoteAddress error variants. The decode
checks cover RemoteAddress, RemoteSupervisor, RemoteObserver,
RemoteSupervisionEvent, the CreateActor/GetActor reply paths in
Session::handle_node_message, and the Send/DoSend dispatch in
Session::handle_actor_message.

+++ acktor-ipc/src/node/command.rs
@@ -24,7 +24,7 @@ pub struct AddActor(pub Recipient<RemoteMessage>);
 /// A command which is used to remove a local actor from a node.
 #[derive(Debug, Message)]
 #[result_type(())]
-pub struct RemoveActor(pub usize);
+pub struct RemoveActor(pub u64);
 
 /// A command which is used by a node to actively connect to another node like a client.
 ///
@@ -73,5 +73,5 @@ pub struct GetRemoteActor {
 #[result_type(Result<RemoteAddress>)]
 pub struct CreateRemoteAddress {
     pub session_label: String,
-    pub actor_index: usize,
+    pub actor_index: u64,
 }
